+++ android/src/main/kotlin/com/example/dengage_flutter/DengageFlutterPlugin.kt
@@ -2,31 +2,30 @@ package com.example.dengage_flutter
 
 import android.app.Activity
 import android.content.Context
+import android.content.Intent
+import android.content.IntentFilter
+import android.util.Log
 import androidx.annotation.NonNull
-import androidx.appcompat.app.AppCompatActivity
 import com.dengage.sdk.DengageEvent
+import com.dengage.sdk.NotificationReceiver
 import com.dengage.sdk.callback.DengageCallback
 import com.dengage.sdk.models.DengageError
 import com.dengage.sdk.models.InboxMessage
+import com.dengage.sdk.models.Message
+import com.dengage.sdk.models.TagItem
 import com.google.gson.Gson
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.embedding.engine.plugins.activity.ActivityAware
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
+import io.flutter.plugin.common.EventChannel
+import io.flutter.plugin.common.EventChannel.EventSink
+import io.flutter.plugin.common.EventChannel.StreamHandler
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
-import io.flutter.plugin.common.EventChannel;
-import io.flutter.plugin.common.EventChannel.EventSink;
-import io.flutter.plugin.common.EventChannel.StreamHandler;
-import android.content.Intent
-import com.dengage.sdk.NotificationReceiver
-import android.content.IntentFilter
-import android.util.Log
-import com.dengage.sdk.models.Message
-import io.flutter.embedding.android.FlutterActivity
-import io.flutter.embedding.engine.FlutterEngine
-import org.json.JSONObject
+import java.util.*
+import kotlin.collections.HashMap
 
 /** DengageFlutterPlugin */
 class DengageFlutterPlugin: FlutterPlugin, MethodCallHandler, DengageResponder(), ActivityAware {
@@ -132,6 +131,8 @@ class DengageFlutterPlugin: FlutterPlugin, MethodCallHandler, DengageResponder()
         this.setNavigation(call, result)
       } else if (call.method == "dEngage#setNavigationWithName") {
         this.setNavigationWithName(call, result)
+      } else if (call.method == "dEngage#setTags") {
+        this.setTags(call, result)
       } else {
         result.notImplemented()
       }
@@ -549,6 +550,43 @@ class DengageFlutterPlugin: FlutterPlugin, MethodCallHandler, DengageResponder()
       DengageCoordinator.sharedInstance.dengageManager!!.setNavigation(appActivity, screenName)
       replySuccess(result, true)
     } catch (ex: Exception) {
+      replyError(result, "error", ex.localizedMessage, ex)
+    }
+  }
+
+  /**
+   * Method to setTags.
+   */
+  private fun setTags (@NonNull call: MethodCall, @NonNull result: Result) {
+    try {
+      val data: List<HashMap<String, Any>> = call.argument("tags")!!
+      Log.d("V/Den/RN/Android", data.toString())
+
+      val finalTags = data.map {
+        if (it["tagName"] != null && it["tagValue"] != null) {
+          if (it["changeTime"] != null && it["changeValue"] != null && it["removeTime"] != null) {
+            TagItem(
+                    it["tagName"] as String,
+                    it["tagValue"] as String,
+                    it["changeTime"] as Date?,
+                    it["changeValue"]?.toString(),
+                    it["removeTime"] as Date?
+            )
+          } else {
+            TagItem(
+                    it["tagName"] as String,
+                    it["tagValue"] as String
+            )
+          }
+        } else {
+          throw Exception("required arugment 'tagName' Or 'tagValue' is missing.")
+        }
+      }
+
+      DengageCoordinator.sharedInstance.dengageManager?.setTags(finalTags)
+      replySuccess(result, true)
+    } catch (ex: Exception) {
+      Log.e("V/Den/RN/:setTagsErr", ex.localizedMessage)
       replyError(result, "error", ex.localizedMessage, ex)
     }
   }